Some more design changes :P

+++ WebCamTest/FingerPrintApplication/Util.cs
@@ -9,18 +9,50 @@ namespace FingerPrintApplication
 {
     class Util
     {
-        public static void CreadeDefaultDataSet()
+        static string configFile = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"FingerScan\fsDs.xml");
+        static string dirPath = (Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"FingerScan"));
+
+        public static bool CreateDefaultDataSet()
         {
-            string configFile = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"FingerScan\fsDs.xml");
-            string dirPath = (Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), @"FingerScan"));
-            if (!File.Exists(configFile))
+            try
             {
-                if (!Directory.Exists(dirPath))
+
+                if (!File.Exists(configFile))
                 {
                     Directory.CreateDirectory(dirPath);
+                    if (File.Exists("fsDs.config"))
+                    {
+                        File.Copy("fsDs.config", configFile);
+                        return true;
+                    }
+                    else
+                    {
+                        return false;
+                    }
                 }
-                File.Copy("fsDs.config", configFile);
+                return true;
             }
+            catch
+            {
+                return false;
+            }
+        }
+
+        public static bool VarifyConfig(FSDataSet fsDataSet)
+        {
+            bool isVarified = false;
+            try
+            {
+                if (fsDataSet != null && fsDataSet.Tables["Users"].Rows.Count > 0)
+                {
+                    isVarified = true;
+                }
+            }
+            catch
+            {
+                isVarified = false;
+            }
+            return isVarified;
         }
 
         public static string GetConfigFilepath()
